Add Book Addition Functionality for One Book
This commit adds javascript for adding one book to the library. It displays title, author, and amount of pages accurately. It also adds a small bit of styling to the cards, such as their grid positions. It fixes as previous problem of the book cards going between the footer and main section. Now, they have been fixed to go into their own container, in the main section. The grid adjusts depending on the window size.